refactor: components utils
